refactor(positions): extract InstrumentGroup to its own file
CODE_GUIDE.md §2 "One primary type per file." `InstrumentGroup` was a
secondary top-level struct in `PositionsTable.swift`; pre-existing on
main, surfaced as a code-review finding during Task 5. The grouping
helper is consumed by both `PositionsTable` (iOS `narrowLayout`'s
`groupContent`) and the macOS Grid extension (`sort.sorted(groups.flatMap(\.rows))`),
so it has independent reuse rights and earned a top-level file.
