+++ Shared/Views/Positions/InstrumentGroup.swift
@@ -1,0 +1,35 @@
+import Foundation
+
+/// Internal grouping helper — splits the rows into Stocks / Crypto / Cash
+/// in spec order. Each group is empty if no row of that kind appears.
+struct InstrumentGroup: Identifiable {
+  enum Kind { case stocks, crypto, cash }
+
+  let kind: Kind
+  let rows: [ValuedPosition]
+  var id: String {
+    switch kind {
+    case .stocks: return "stocks"
+    case .crypto: return "crypto"
+    case .cash: return "cash"
+    }
+  }
+  var title: String {
+    switch kind {
+    case .stocks: return "Stocks"
+    case .crypto: return "Crypto"
+    case .cash: return "Cash"
+    }
+  }
+
+  static func from(_ rows: [ValuedPosition]) -> [InstrumentGroup] {
+    let stocks = rows.filter { $0.instrument.kind == .stock }
+    let crypto = rows.filter { $0.instrument.kind == .cryptoToken }
+    let cash = rows.filter { $0.instrument.kind == .fiatCurrency }
+    return [
+      .init(kind: .stocks, rows: stocks),
+      .init(kind: .crypto, rows: crypto),
+      .init(kind: .cash, rows: cash),
+    ].filter { !$0.rows.isEmpty }
+  }
+}
