+++ Shared/Views/Positions/PositionsTable.swift
@@ -276,40 +276,6 @@ extension PositionsTable {
   }
 }
 
-/// Internal grouping helper — splits the rows into Stocks / Crypto / Cash
-/// in spec order. Each group is empty if no row of that kind appears.
-struct InstrumentGroup: Identifiable {
-  enum Kind { case stocks, crypto, cash }
-
-  let kind: Kind
-  let rows: [ValuedPosition]
-  var id: String {
-    switch kind {
-    case .stocks: return "stocks"
-    case .crypto: return "crypto"
-    case .cash: return "cash"
-    }
-  }
-  var title: String {
-    switch kind {
-    case .stocks: return "Stocks"
-    case .crypto: return "Crypto"
-    case .cash: return "Cash"
-    }
-  }
-
-  static func from(_ rows: [ValuedPosition]) -> [InstrumentGroup] {
-    let stocks = rows.filter { $0.instrument.kind == .stock }
-    let crypto = rows.filter { $0.instrument.kind == .cryptoToken }
-    let cash = rows.filter { $0.instrument.kind == .fiatCurrency }
-    return [
-      .init(kind: .stocks, rows: stocks),
-      .init(kind: .crypto, rows: crypto),
-      .init(kind: .cash, rows: cash),
-    ].filter { !$0.rows.isEmpty }
-  }
-}
-
 private func mixedPositionsInput() -> PositionsViewInput {
   let bhp = Instrument.stock(ticker: "BHP.AX", exchange: "ASX", name: "BHP")
   let cba = Instrument.stock(ticker: "CBA.AX", exchange: "ASX", name: "CBA")
